Handle different worker-user-data secret name for 4.10
The worker-user-data secret name became worker-user-data-managed in
4.10. Hive now supports alternate names for this secret.

When the OpenShift version reported in the ClusterDeployment cannot
be parsed as a semver, then assume that the cluster is using the
new name for the worker user data secret. We know that legacy
OpenShift versions use semver, so we can conclude that any version
that is not using semver must not be a legacy version.

This is a cherry-pick (updated to start with 4.10 instead of 4.6) of the
original patch 4b98f3ae342f4600cc1225f2a891cbf8e78e8abb, which was later
reverted by 45cabd4f984d8fdc3114ae68e7176e7c7fa36985.